update with new structure
